refactor: 레디스 캐싱 코드 정리
